Finally stop blocking UI
This one is kind of funny so I'll put it here:
After failing to make the functions work asynchronously to display search results, I couldn't figure out why it was blocking my application for so long.

I decided to run the SQL query by itself, and lo and behold it was nearly instant (as expected). However, it returned 3000 results! I realized that I was creating over 3000 bitmaps and Images and not limiting my results...

Anyway it works now.